Bug Fixes, translations, fix display/instructions.
Fixed race condition on ccda button generator if DOM ready event wasn't
fired.

Removed the menu hiding condition if the smart app config was not setup.
This makes it so people can see the instructions for setup.

Removed unneeded scopes in the setup instructions.

Fixed language translation constants missing.

Added the patient name to be displayed as part of the authorization.